requirements.txt
fastapi
uvicorn
google-generativeai
python-dotenv
sqlalchemy
faiss-cpu
sentence-transformers
pyjwt
passlib[bcrypt]
python-multipart
pydantic
requests
